Перенес экраны и фасады,связанные со словарями в package dictionaries
