--- app/src/main/kotlin/com/designdrivendevelopment/kotelok/PlaySoundBtnClickListener.kt
@@ -1,5 +1,0 @@
-package com.designdrivendevelopment.kotelok
-
-interface PlaySoundBtnClickListener {
-    fun onPlayBtnClick(text: String)
-}
