+++ app/src/main/kotlin/com/designdrivendevelopment/kotelok/screens/dictionaries/EditWordDefinitionsRepository.kt
@@ -1,5 +1,6 @@
-package com.designdrivendevelopment.kotelok
+package com.designdrivendevelopment.kotelok.screens.dictionaries
 
+import com.designdrivendevelopment.kotelok.DefinitionsRequestResult
 import com.designdrivendevelopment.kotelok.entities.Dictionary
 import com.designdrivendevelopment.kotelok.entities.WordDefinition
 import kotlinx.coroutines.flow.Flow
