+++ app/src/main/kotlin/com/designdrivendevelopment/kotelok/screens/dictionaries/dictionaryDetailsScreen/PlaySoundBtnClickListener.kt
@@ -1,0 +1,5 @@
+package com.designdrivendevelopment.kotelok.screens.dictionaries.dictionaryDetailsScreen
+
+interface PlaySoundBtnClickListener {
+    fun onPlayBtnClick(text: String)
+}
