+++ app/src/main/kotlin/com/designdrivendevelopment/kotelok/screens/dictionaries/lookupWordDefinitionsScreen/viewTypes/CommonItem.kt
@@ -1,0 +1,5 @@
+package com.designdrivendevelopment.kotelok.screens.dictionaries.lookupWordDefinitionsScreen.viewTypes
+
+abstract class CommonItem<out T : Any?> : ItemWithType {
+    abstract val data: T
+}
