+++ app/src/main/kotlin/com/designdrivendevelopment/kotelok/screens/dictionaries/lookupWordDefinitionsScreen/viewTypes/ItemWithType.kt
@@ -1,0 +1,5 @@
+package com.designdrivendevelopment.kotelok.screens.dictionaries.lookupWordDefinitionsScreen.viewTypes
+
+interface ItemWithType {
+    val viewType: Int
+}
