--- app/src/main/kotlin/com/designdrivendevelopment/kotelok/screens/wordDefinitions/lookupWordDefinitionsScreen/viewTypes/CommonItem.kt
@@ -1,5 +1,0 @@
-package com.designdrivendevelopment.kotelok.screens.wordDefinitions.lookupWordDefinitionsScreen.viewTypes
-
-abstract class CommonItem<out T : Any?> : ItemWithType {
-    abstract val data: T
-}
