--- app/src/main/kotlin/com/designdrivendevelopment/kotelok/screens/wordDefinitions/lookupWordDefinitionsScreen/viewTypes/ItemWithType.kt
@@ -1,5 +1,0 @@
-package com.designdrivendevelopment.kotelok.screens.wordDefinitions.lookupWordDefinitionsScreen.viewTypes
-
-interface ItemWithType {
-    val viewType: Int
-}
