+++ app/src/main/res/layout/fragment_dictionaries.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".screens.dictionaries.DictionariesFragment">
+    tools:context=".screens.dictionaries.dictionariesScreen.DictionariesFragment">
 
     <TextView
         android:id="@+id/dictionary_text"
