+++ app/src/main/res/layout/fragment_dictionary_details.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".DictionaryDetailsFragment">
+    tools:context=".screens.dictionaries.dictionaryDetailsScreen.DictionaryDetailsFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/word_definitions_list"
