+++ app/src/main/res/layout/fragment_lookup_word_definition.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:background="?attr/colorOnPrimary"
-    tools:context=".screens.wordDefinitions.lookupWordDefinitionsScreen.LookupWordDefinitionsFragment">
+    tools:context=".screens.dictionaries.lookupWordDefinitionsScreen.LookupWordDefinitionsFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
